:sparkles: Feat: front css 적용 및 back 버그 수정

See merge request s08-webmobile1-sub2/S08P12A702!150

+++ back/Dream-Hi/src/main/java/com/elephant/dreamhi/service/MediaFileService.java
@@ -9,6 +9,7 @@ import com.elephant.dreamhi.model.statics.MediaType;
 import com.elephant.dreamhi.repository.ActorProfileMediaFileRepository;
 import com.elephant.dreamhi.repository.ActorRepository;
 import java.util.List;
+import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
@@ -24,9 +25,10 @@ import org.springframework.transaction.annotation.Transactional;
 public class MediaFileService {
 
     @Value("${app.image-size}")
-    private Integer imageSize;
+    private static Integer imageSize;
     @Value("${app.video-size}")
-    private Integer videoSize;
+    private static Integer videoSize;
+
     private final ActorProfileMediaFileRepository actorProfileMediaFileRepository;
     private final ActorRepository actorRepository;
 
@@ -45,15 +47,18 @@ public class MediaFileService {
      * @throws FullResourceException : 저장 공간 부족
      */
     @Transactional
-    public void addMediaFile(Long userId, Long actorProfileId, MediaFileRequestDto mediaFileRequestDto)
+    public Long addMediaFile(Long userId, Long actorProfileId, MediaFileRequestDto mediaFileRequestDto)
             throws AccessDeniedException, FullResourceException {
         ActorProfile actorProfile = actorRepository.findByIdAndUser_Id(actorProfileId, userId)
-                                                   .orElseThrow(() -> new AccessDeniedException(userId +" 유저는 " + actorProfileId + " 프로필 수정 권한이 없습니다."));
+                                                   .orElseThrow(
+                                                           () -> new AccessDeniedException(userId + " 유저는 " + actorProfileId + " 프로필 수정 권한이 없습니다."));
 
         checkStorageSize(mediaFileRequestDto, actorProfile);
 
         ActorProfileMediaFile mediaFile = mediaFileRequestDto.toEntity();
         mediaFile.changeActorProfile(actorProfile);
+
+        return actorProfile.getId();
     }
 
 
@@ -69,7 +74,12 @@ public class MediaFileService {
     @Transactional
     public void deleteMediaFile(Long userId, Long actorProfileId, Long actorProfileMediaFileId)
             throws AccessDeniedException, EmptyResultDataAccessException {
-        actorRepository.checkValidateModify(actorProfileId, userId).orElseThrow(() -> new AccessDeniedException(userId + " 유저는 " + actorProfileId + "번 프로필 삭제 권한이 없습니다."));
+        Optional<ActorProfile> actorProfile = actorRepository.checkValidateModify(actorProfileId, userId);
+
+        if (actorProfile.isEmpty()) {
+            throw new AccessDeniedException(userId + " 유저는 " + actorProfileId + "번 프로필 삭제 권한이 없습니다.");
+        }
+
         actorProfileMediaFileRepository.deleteById(actorProfileMediaFileId);
     }
